+++ src/containers/subjects/list.js
@@ -4,67 +4,133 @@ import { bindActionCreators } from 'redux';
 import { View, Alert } from 'react-native';
 import SubjectsListComponent from '../../components/subjects/list';
 import { subjectsSelector } from '../../redux/selectors';
-import { deleteSubject } from '../../redux/actions';
+import { deleteSubjects } from '../../redux/actions';
 import { NewButton } from '../../shared/UI/buttons';
+import withItemSelection from '../../hoc/itemSelection';
 
 class SubjectsList extends React.Component {
+  static getSelectionActions(navigation, amountSelected) {
+    return [
+      {
+        enabled: amountSelected === 1,
+        icon: 'edit',
+        handlePress: navigation.getParam('handleEditSelected'),
+      },
+      {
+        enabled: true,
+        icon: 'delete',
+        handlePress: navigation.getParam('handleDeleteSelected'),
+      },
+    ];
+  }
+
+  static navigationOptions() {
+    return {
+      title: 'Subjects', // DICTIONARY
+    };
+  }
+
   constructor(props) {
     super(props);
 
-    this.findSubject = this.findSubject.bind(this);
-    this.handlePressDetailSubject = this.handlePressDetailSubject.bind(this);
-    this.handlePressEditSubject = this.handlePressEditSubject.bind(this);
-    this.handlePressDeleteSubject = this.handlePressDeleteSubject.bind(this);
+    this.handleLongPressSubject = this.handleLongPressSubject.bind(this);
+    this.handlePressSubject = this.handlePressSubject.bind(this);
+
     this.handlePressNewSubject = this.handlePressNewSubject.bind(this);
+    this.handleEditSelected = this.handleEditSelected.bind(this);
+    this.handleDeleteSelected = this.handleDeleteSelected.bind(this);
+
+    this.props.navigation.setParams({
+      handleEditSelected: this.handleEditSelected,
+      handleDeleteSelected: this.handleDeleteSelected,
+    });
   }
 
   findSubject(id) {
-    return this.props.subjects.find(subj => subj.id === id);
+    const numericId = Number(id);
+    return this.props.subjects.find(subj => subj.id === numericId);
   }
 
-  handlePressDetailSubject(id) {
-    const subject = this.findSubject(id);
-    if (subject) {
-      this.props.navigation.navigate('showSubject', { subject });
+  handlePressSubject(id) {
+    const { amountSelected } = this.props;
+    if (amountSelected) {
+      this.props.toggleSelection(id);
+    } else {
+      const subject = this.findSubject(id);
+      if (subject) {
+        this.props.navigation.navigate('showSubject', { subject });
+      }
     }
   }
 
-  handlePressEditSubject(id) {
-    const subject = this.findSubject(id);
+  handleLongPressSubject(id) {
+    this.props.toggleSelection(id);
+  }
+
+  handlePressNewSubject() {
+    if (this.props.amountSelected) {
+      return;
+    }
+    this.props.navigation.navigate('newSubject');
+  }
+
+  handleEditSelected() {
+    const selectedIds = this.props.getSelectionArray();
+    if (selectedIds.length !== 1) {
+      return;
+    }
+
+    const subject = this.findSubject(selectedIds[0]);
     if (subject) {
+      this.props.clearSelection();
       this.props.navigation.navigate('editSubject', { subject });
     }
   }
 
-  handlePressDeleteSubject(id) {
-    const subject = this.findSubject(id);
+  handleDeleteSelected() {
+    const selectedIds = this.props.getSelectionArray();
+    if (selectedIds.length === 0) {
+      return;
+    }
+
+    let deletionSelected;
+    if (selectedIds.length === 1) {
+      const subject = this.findSubject(selectedIds[0]);
+      deletionSelected = subject.name;
+    } else {
+      deletionSelected = `${selectedIds.length} subjects`;
+    }
+
     // DICTIONARY
     Alert.alert(
       'Confirmation',
-      `Are you sure you want to delete ${subject.name}?`,
+      `Delete ${deletionSelected}?`,
       [
         { text: 'Cancel', style: 'cancel' },
-        { text: 'Delete', onPress: () => this.props.deleteSubject(id) },
+        {
+          text: 'Delete',
+          onPress: () => {
+            this.props.deleteSubjects(selectedIds);
+            this.props.clearSelection();
+          },
+        },
       ],
     );
   }
 
-  handlePressNewSubject() {
-    this.props.navigation.navigate('newSubject');
-  }
-
   render() {
-    const { subjects } = this.props;
+    const { subjects, selection, amountSelected } = this.props;
 
     return (
       <View style={{ flex: 1 }}>
         <SubjectsListComponent
           subjects={subjects}
-          onPressDetail={this.handlePressDetailSubject}
-          onPressEdit={this.handlePressEditSubject}
-          onPressDelete={this.handlePressDeleteSubject}
+          selectedSubjects={selection}
+          onPressSubject={this.handlePressSubject}
+          onLongPressSubject={this.handleLongPressSubject}
         />
         <NewButton
+          disabled={amountSelected > 0}
           onPress={this.handlePressNewSubject}
         />
       </View>
@@ -77,7 +143,9 @@ const mapStateToProps = state => ({
 });
 
 const mapDispatchToProps = dispatch => bindActionCreators({
-  deleteSubject,
+  deleteSubjects,
 }, dispatch);
 
-export default connect(mapStateToProps, mapDispatchToProps)(SubjectsList);
+export default withItemSelection(connect(
+  mapStateToProps, mapDispatchToProps,
+)(SubjectsList));