+++ src/shared/UI/icons/index.js
@@ -1,5 +1,11 @@
+import React from 'react';
 import HamburgerIcon from './hamburger';
+import ClickableIcon from './clickable';
+
+const CrossIcon = props => <ClickableIcon {...props} icon="cross" />;
 
 export {
-  HamburgerIcon, // eslint-disable-line import/prefer-default-export
+  CrossIcon,
+  HamburgerIcon,
+  ClickableIcon,
 };